Fikset noen exports, flyttet knappe-valg in i egen komponenet. Slettet SettPåVentEnkel -> som er nå SettPåVent.tsx.
