+++ src/frontend/Komponenter/Behandling/BehandlingContainer.tsx
@@ -14,7 +14,7 @@ import { HenleggModal } from './Henleggelse/HenleggModal';
 import ScrollToTop from '../../Felles/ScrollToTop/ScrollToTop';
 import { useSetPersonIdent } from '../../App/hooks/useSetPersonIdent';
 import { useSetValgtFagsakId } from '../../App/hooks/useSetValgtFagsakId';
-import { SettPåVentEnkel } from './SettPåVent/SettPåVentEnkel';
+import SettPåVent from './SettPåVent/SettPåVent';
 
 const Container = styled.div`
     display: flex;
@@ -73,7 +73,7 @@ const BehandlingContent: FC<{
             <Container>
                 <InnholdWrapper åpenHøyremeny={åpenHøyremeny}>
                     <Fanemeny behandling={behandling} />
-                    <SettPåVentEnkel behandling={behandling}></SettPåVentEnkel>
+                    <SettPåVent behandling={behandling}></SettPåVent>
                     <BehandlingRoutes behandling={behandling} />
                     <HenleggModal behandling={behandling} />
                 </InnholdWrapper>
